sport game controller done

+++ src/main/java/com/larrex/BettingTicketBookingSystem/controller/SportCategoryController.java
@@ -4,8 +4,10 @@ import com.larrex.BettingTicketBookingSystem.model.SportCategory;
 import com.larrex.BettingTicketBookingSystem.model.Status;
 import com.larrex.BettingTicketBookingSystem.repository.SportCategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.UUID;
@@ -29,24 +31,42 @@ public class SportCategoryController {
         return ResponseEntity.ok(repository.findAll());
     }
 
+    @GetMapping("get/{id}")
+    public ResponseEntity<SportCategory> getSportCategoryWithId(@PathVariable String id) {
+
+        SportCategory sportCategory = repository.getSingle(id);
+
+        if (sportCategory != null)
+            return ResponseEntity.ok(sportCategory);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No sport category found with id " + id);
+    }
+
     @PutMapping("update/{id}")
-    public ResponseEntity<Status> updateSPortCategoryWithId(@PathVariable String id, @RequestBody SportCategory oldSportCategory) {
+    public ResponseEntity<Status> updateSportCategoryWithId(@PathVariable String id, @RequestBody SportCategory oldSportCategory) {
 
         SportCategory newSportCategory = repository.getSingle(id);
         newSportCategory.setCategoryName(oldSportCategory.getCategoryName());
 
-        repository.save(newSportCategory);
+        SportCategory justSavedSportCategory = repository.save(newSportCategory);
 
-        return ResponseEntity.ok(Status.SUCCESS);
+        if (justSavedSportCategory != null)
+            return ResponseEntity.ok(Status.SUCCESS);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No sport category found with id " + id);
     }
 
     @DeleteMapping("delete/{id}")
     public ResponseEntity<Status> deleteSportCategoryWithId(@PathVariable String id) {
 
         SportCategory sportCategory = repository.getSingle(id);
+        if (sportCategory != null) {
+            repository.delete(sportCategory);
+            return ResponseEntity.ok(Status.SUCCESS);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No sport category found with id " + id);
+        }
 
-        repository.delete(sportCategory);
-        return ResponseEntity.ok(Status.SUCCESS);
     }
 
 }
